finished category creation and dropdown for inventory creation

+++ RektaRetailApp.Domain/Abstractions/BaseIdentityUser.cs
@@ -5,7 +5,7 @@ using Microsoft.AspNetCore.Identity;
 
 namespace RektaRetailApp.Domain.Abstractions
 {
-    public class BaseIdentityUser : IdentityUser<long>, IDomainModel
+    public class BaseIdentityUser : IdentityUser<int>, IDomainModel
     {
         public DateTimeOffset CreatedAt { get; set; }
         public DateTimeOffset UpdatedAt { get; set; }
